refact: extract and reuse run once hook
