+++ packages/designmanual/stories/article/AudioExample.jsx
@@ -1,66 +1,48 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
+import React from 'react';
 import { AudioPlayer, Figure } from '@ndla/ui';
 import { uuid } from '@ndla/util';
 import { getLicenseByAbbreviation } from '@ndla/licenses';
-import {
-  addCloseDialogClickListeners,
-  addShowDialogClickListeners,
-  initAudioPlayers,
-} from '@ndla/article-scripts';
+import { initArticleScripts } from '@ndla/article-scripts';
 import { FigureCaptionExample } from './FigureCaptionExample';
+import { useRunOnlyOnce } from './useRunOnlyOnce';
 
-class AudioExample extends Component {
-  constructor(props) {
-    super(props);
-    this.id = uuid();
-  }
+function AudioExample() {
+  const id = useRunOnlyOnce(uuid(), () => {
+    initArticleScripts();
+  });
+  const license = getLicenseByAbbreviation('CC-BY-ND-4.0', 'nb');
 
-  componentDidMount() {
-    if (this.props.runScripts) {
-      addShowDialogClickListeners();
-      addCloseDialogClickListeners();
-      initAudioPlayers();
-    }
-  }
+  const messages = {
+    close: 'Lukk',
+    rulesForUse: 'Regler for bruk av lydklippet',
+    learnAboutLicenses: license.linkText,
+    source: 'Kilde',
+    title: 'Tittel',
+    mediaType: 'lydklipp',
+  };
 
-  render() {
-    const license = getLicenseByAbbreviation('CC-BY-ND-4.0', 'nb');
+  const caption = 'Familien som spela vekk jula';
 
-    const messages = {
-      close: 'Lukk',
-      rulesForUse: 'Regler for bruk av lydklippet',
-      learnAboutLicenses: license.linkText,
-      source: 'Kilde',
-      title: 'Tittel',
-      mediaType: 'lydklipp',
-    };
+  const figureId = `figure-${id}`;
 
-    const caption = 'Familien som spela vekk jula';
-
-    const figureId = `figure-${this.id}`;
-
-    return (
-      <Figure id={figureId} type="full-column">
-        <AudioPlayer
-          src="https://staging.api.ndla.no/audio/files/Alltid_Nyheter_nrk128kps.mp3"
-          type="audio/mpeg"
-          title={caption}
-          typeLabel="Hørespill"
-        />
-        <FigureCaptionExample
-          caption={caption}
-          figureId={figureId}
-          id={this.id}
-          messages={messages}
-        />
-      </Figure>
-    );
-  }
+  return (
+    <Figure id={figureId} type="full-column">
+      <AudioPlayer
+        src="https://staging.api.ndla.no/audio/files/Alltid_Nyheter_nrk128kps.mp3"
+        type="audio/mpeg"
+        title={caption}
+        typeLabel="Hørespill"
+      />
+      <FigureCaptionExample
+        caption={caption}
+        figureId={figureId}
+        id={id}
+        messages={messages}
+      />
+    </Figure>
+  );
 }
 
-AudioExample.propTypes = {
-  runScripts: PropTypes.bool,
-};
+AudioExample.propTypes = {};
 
 export default AudioExample;
