+++ packages/designmanual/stories/article/FigureWithLicense.jsx
@@ -5,71 +5,49 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import React, { Component } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import { uuid } from '@ndla/util';
-import {
-  addCloseDialogClickListeners,
-  addShowDialogClickListeners,
-  addEventListenerForResize,
-  updateIFrameDimensions,
-  toggleLicenseInfoBox,
-  addEventListenerForFigureZoomButton,
-} from '@ndla/article-scripts';
+import { initArticleScripts } from '@ndla/article-scripts';
 
 import { Figure } from '@ndla/ui';
 import { FigureCaptionExample } from './FigureCaptionExample';
+import { useRunOnlyOnce } from './useRunOnlyOnce';
 
-class FigureWithLicense extends Component {
-  constructor(props) {
-    super(props);
-    this.id = uuid();
-  }
+function FigureWithLicense({
+  children,
+  hasHiddenCaption,
+  messages,
+  resizeIframe,
+  caption,
+  type,
+}) {
+  const id = useRunOnlyOnce(uuid(), () => {
+    initArticleScripts();
+  });
 
-  componentDidMount() {
-    if (this.props.runScripts) {
-      addShowDialogClickListeners();
-      addCloseDialogClickListeners();
-      updateIFrameDimensions();
-      addEventListenerForResize();
-      toggleLicenseInfoBox();
-      addEventListenerForFigureZoomButton();
-    }
-  }
+  const figureId = `figure-${id}`;
 
-  render() {
-    const {
-      hasHiddenCaption,
-      messages,
-      resizeIframe,
-      caption,
-      type,
-    } = this.props;
-
-    const figureId = `figure-${this.id}`;
-
-    return (
-      <Figure id={figureId} resizeIframe={resizeIframe} type={type}>
-        {this.props.children}
-        {!hasHiddenCaption && (
-          <FigureCaptionExample
-            id={this.id}
-            figureId={figureId}
-            caption={caption}
-            messages={messages}
-            hasHiddenCaption={hasHiddenCaption}
-          />
-        )}
-      </Figure>
-    );
-  }
+  return (
+    <Figure id={figureId} resizeIframe={resizeIframe} type={type}>
+      {children}
+      {!hasHiddenCaption && (
+        <FigureCaptionExample
+          id={id}
+          figureId={figureId}
+          caption={caption}
+          messages={messages}
+          hasHiddenCaption={hasHiddenCaption}
+        />
+      )}
+    </Figure>
+  );
 }
 
 FigureWithLicense.propTypes = {
   children: PropTypes.node.isRequired,
   caption: PropTypes.string,
   reuseLabel: PropTypes.string,
-  runScripts: PropTypes.bool,
   resizeIframe: PropTypes.bool,
   hasHiddenCaption: PropTypes.bool,
   messages: PropTypes.object,
@@ -87,7 +65,6 @@ FigureWithLicense.propTypes = {
 
 FigureWithLicense.defaultProps = {
   caption: '',
-  runScripts: false,
   noFigcaption: false,
   hasHiddenCaption: false,
 };
